flexible cache limit
Cache limit can now be either plain item count or the memory footprint.
The type of limit can be specified (as a function) along with the limit value.
Two strategies are already supplied, `maxcount` and `maxmem`.

The `maxmem` strategy uses `Base.symmarysize` to estimate the memory footprint of cached items.